requirements.txt
fastapi~=0.115.4
uvicorn
pandas~=2.2.3
scikit-learn~=1.5.2
joblib~=1.4.2
pydantic~=2.9.2
requests~=2.32.3
numpy~=2.1.3
app~=0.0.1
Routes~=2.5.1
scikit-surprise~=1.1.4
python-dotenv~=1.0.1